+++ netlify/patch-files.sh
@@ -7,8 +7,8 @@ set -e
 
 echo 'Patch intro.md to include netlify banner.'
 cp netlify/intro.md public/mock-backend/public/intro.md
-echo 'Patch motd.txt to include privacy policy.'
-cp netlify/motd.txt public/mock-backend/public/motd.txt
+echo 'Patch motd.md to include privacy policy.'
+cp netlify/motd.md public/mock-backend/public/motd.md
 echo 'Patch version.json to include git hash'
 jq ".version = \"0.0.0+${GITHUB_SHA:0:8}\"" src/version.json > src/_version.json
 mv src/_version.json src/version.json